requirements.txt
mcp
click>=8.0.0
boto3
botocore[crt]
rich>=13.0.0
python-dotenv
strands-agents
strands-agents-tools
markdown
python-docx
python-pptx
